+++ src/components/ EntranceButton/index.tsx
