+++ src/components/EntranceLabel/index.tsx
